Added an actual Unsplash API app ID to sample app.
It's included into the manifest via manifest placeholders in the gradle file. The actual ID
is stored in sample/gradle.properties, which I've excluded from the repo for security
reasons (not sure if the app ID needs to be kept secret, but I figure I might as well do
so).

I also update the Sample scope to include the build.gradle and other files at the top level
of the sample module.